src/python/requirements.txt
Flask==2.2.2
prometheus-client==0.14.1
PyGithub==1.57
pyjson==1.3.0
PyYAML==6.0
requests==2.22.0
APScheduler==3.9.1
